// src/redux/createEntityAttestation/createEntityAttestation.types.ts
export interface ClaimInfo {
  title: string
  shortDescription: string
  // type: EntityClaimType
  type: string
  feature?: string
  reliability?: string
  userGuide?: string
  reference?: string
  keywords?: string[]
}

export interface Question {
  id: string
  title: string
  description: string
  label: string
  required: boolean
  type: Type
  control: ControlType
  attributeType: string
  minItems?: number
  maxItems?: number
  values?: string[]
  itemValues?: string[]
  itemLabels?: string[]
  itemImages?: string[]
  placeholder?: string
  initialValue?: string
  inline?: boolean
  order: number
  currency?: string
}

export interface CreateEntityAttestationState {
  claimInfo: ClaimInfo
  questions: {
    [id: string]: Question
  }
  validation: {
    [identifier: string]: Validation
  }
}